break out all the tiny functions
