+++ src/functions/clean.js
@@ -1,0 +1,6 @@
+var parse = require('./parse');
+
+module.exports = function clean (version, options) {
+  var s = parse(version.trim().replace(/^[=v]+/, ''), options)
+  return s ? s.version : null
+}
