+++ src/functions/compare.js
@@ -1,0 +1,5 @@
+var SemVer = require('../SemVer');
+
+module.exports = function compare (a, b, loose) {
+  return new SemVer(a, loose).compare(new SemVer(b, loose))
+}
