+++ src/functions/compareBuild.js
@@ -1,0 +1,7 @@
+var SemVer = require('../SemVer');
+
+module.exports = function compareBuild (a, b, loose) {
+  var versionA = new SemVer(a, loose)
+  var versionB = new SemVer(b, loose)
+  return versionA.compare(versionB) || versionA.compareBuild(versionB)
+}
