+++ src/functions/compareLoose.js
@@ -1,0 +1,5 @@
+var compare = require('./compare');
+
+module.exports = function compareLoose (a, b) {
+  return compare(a, b, true)
+}
