+++ src/functions/eq.js
@@ -1,0 +1,5 @@
+var compare = require('./compare');
+
+module.exports = function eq (a, b, loose) {
+  return compare(a, b, loose) === 0
+}
