+++ src/functions/gt.js
@@ -1,0 +1,5 @@
+var compare = require('./compare');
+
+module.exports = function gt (a, b, loose) {
+  return compare(a, b, loose) > 0
+}
