+++ src/functions/gte.js
@@ -1,0 +1,6 @@
+
+var compare = require('./compare');
+
+module.exports = function gte (a, b, loose) {
+  return compare(a, b, loose) >= 0
+}
