+++ src/functions/inc.js
@@ -1,0 +1,14 @@
+var SemVer = require('../SemVer');
+
+module.exports = function inc (version, release, loose, identifier) {
+  if (typeof (loose) === 'string') {
+    identifier = loose
+    loose = undefined
+  }
+
+  try {
+    return new SemVer(version, loose).inc(release, identifier).version
+  } catch (er) {
+    return null
+  }
+}
