+++ src/functions/major.js
@@ -1,0 +1,5 @@
+var SemVer = require('../SemVer');
+
+module.exports = function major (a, loose) {
+  return new SemVer(a, loose).major
+}
