+++ src/functions/minor.js
@@ -1,0 +1,5 @@
+var SemVer = require('../SemVer');
+
+module.exports = function minor (a, loose) {
+  return new SemVer(a, loose).minor
+}
