+++ src/functions/neq.js
@@ -1,0 +1,5 @@
+var compare = require('./compare');
+
+module.exports = function neq (a, b, loose) {
+  return compare(a, b, loose) !== 0
+}
