+++ src/functions/patch.js
@@ -1,0 +1,5 @@
+var SemVer = require('../SemVer');
+
+module.exports = function patch (a, loose) {
+  return new SemVer(a, loose).patch
+}
