+++ src/functions/prerelease.js
@@ -1,0 +1,6 @@
+var parse = require('./parse');
+
+module.exports =function prerelease (version, options) {
+  var parsed = parse(version, options)
+  return (parsed && parsed.prerelease.length) ? parsed.prerelease : null
+}
