+++ src/functions/rcompare.js
@@ -1,0 +1,5 @@
+var compare = require('./compare');
+
+module.exports = function rcompare (a, b, loose) {
+  return compare(b, a, loose)
+}
