+++ src/functions/rsort.js
@@ -1,0 +1,7 @@
+var compareBuild = require('./compareBuild');
+
+module.exports = function rsort (list, loose) {
+    return list.sort(function (a, b) {
+      return compareBuild(b, a, loose)
+    })
+  }
