+++ src/functions/sort.js
@@ -1,0 +1,7 @@
+var compareBuild = require('./compareBuild');
+
+module.exports = function sort (list, loose) {
+    return list.sort(function (a, b) {
+      return compareBuild(a, b, loose)
+    })
+  }
