+++ src/functions/valid.js
@@ -1,0 +1,6 @@
+var parse = require('./parse');
+
+module.exports = function valid (version, options) {
+  var v = parse(version, options)
+  return v ? v.version : null
+};
