+++ src/ranges/toComparators.js
@@ -1,0 +1,10 @@
+var { Range } = require('./index');
+
+// Mostly just for testing and legacy API reasons
+module.exports = function toComparators (range, options) {
+  return new Range(range, options).set.map(function (comp) {
+    return comp.map(function (c) {
+      return c.value
+    }).join(' ').trim().split(' ')
+  })
+}
