+++ src/ranges/validRange.js
@@ -1,0 +1,10 @@
+var { Range } = require('../index');
+module.exports = function validRange (range, options) {
+    try {
+      // Return '*' instead of '' so that truthiness works.
+      // This will throw if it's invalid anyway
+      return new Range(range, options).range || '*'
+    } catch (er) {
+      return null
+    }
+  }
